Update fuzzer to run iterate
